+++ pages/_app.js
@@ -1,10 +1,12 @@
 /* eslint filenames/match-exported: 0 */
 /* eslint react/prop-types: 0 */
 
-import { StorageProvider } from 'context/storage';
 import { DAppProvider } from '@usedapp/core';
 import { ToastContainer } from 'react-toastify';
 import { ReactQueryDevtools } from 'react-query/devtools';
+
+import { StorageProvider } from 'context/storage';
+import { UnlockProvider } from 'context/unlock';
 
 import { QueryClient, QueryClientProvider } from 'react-query';
 import 'react-toastify/dist/ReactToastify.css';
@@ -19,11 +21,13 @@ function App({ Component, pageProps }) {
   return (
     <QueryClientProvider client={queryClient}>
       <DAppProvider config={config}>
-        <StorageProvider>
-          <Component {...pageProps} />
-          <ToastContainer />
-          <ReactQueryDevtools initialIsOpen={false} position="bottom-right" />
-        </StorageProvider>
+        <UnlockProvider>
+          <StorageProvider>
+            <Component {...pageProps} />
+            <ToastContainer />
+            <ReactQueryDevtools initialIsOpen={false} position="bottom-right" />
+          </StorageProvider>
+        </UnlockProvider>
       </DAppProvider>
     </QueryClientProvider>
   );